refactor: separate spatial scene into different files
