+++ src/component/spatial/enums/CameraVisualizationMode.ts
