+++ src/component/spatial/enums/OriginalPositionMode.ts
